+++ bgce-timetracker/Controllers/UsersController.cs
@@ -17,34 +17,55 @@ namespace bgce_timetracker.Controllers
         // GET: Users
         public ActionResult Index()
         {
-            var uSERs = db.USERs.Include(u => u.LOCATION1).Include(u => u.PAID_STAFF).Include(u => u.UNIT_DIRECTOR).Include(u => u.USER2).Include(u => u.VOLUNTEER);
-            return View(uSERs.ToList());
+            if (Request.IsAuthenticated)
+            {
+                var uSERs = db.USERs.Include(u => u.LOCATION1).Include(u => u.PAID_STAFF).Include(u => u.UNIT_DIRECTOR).Include(u => u.USER2).Include(u => u.VOLUNTEER);
+                return View(uSERs.ToList());
+            }
+            else
+            {
+                return RedirectToAction("Index", "Home");
+            }
         }
 
         // GET: Users/Details/5
         public ActionResult Details(int? id)
         {
-            if (id == null)
+            if (Request.IsAuthenticated)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                USER uSER = db.USERs.Find(id);
+                if (uSER == null)
+                {
+                    return HttpNotFound();
+                }
+                return View(uSER);
             }
-            USER uSER = db.USERs.Find(id);
-            if (uSER == null)
+            else
             {
-                return HttpNotFound();
+                return RedirectToAction("Index", "Home");
             }
-            return View(uSER);
         }
 
         // GET: Users/Create
         public ActionResult Create()
         {
-            ViewBag.location = new SelectList(db.LOCATIONs, "locationID", "name");
-            ViewBag.userID = new SelectList(db.PAID_STAFF, "emplID", "pay_schedule");
-            ViewBag.userID = new SelectList(db.UNIT_DIRECTOR, "emplID", "emplID");
-            ViewBag.manager = new SelectList(db.USERs, "userID", "fname");
-            ViewBag.userID = new SelectList(db.VOLUNTEERs, "volID", "volID");
-            return View();
+            if (Request.IsAuthenticated)
+            {
+                ViewBag.location = new SelectList(db.LOCATIONs, "locationID", "name");
+                ViewBag.userID = new SelectList(db.PAID_STAFF, "emplID", "pay_schedule");
+                ViewBag.userID = new SelectList(db.UNIT_DIRECTOR, "emplID", "emplID");
+                ViewBag.manager = new SelectList(db.USERs, "userID", "fname");
+                ViewBag.userID = new SelectList(db.VOLUNTEERs, "volID", "volID");
+                return View();
+            }
+            else
+            {
+                return RedirectToAction("Index", "Home");
+            }
         }
 
         // POST: Users/Create
@@ -54,51 +75,65 @@ namespace bgce_timetracker.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Create(USER newUser)
         {
-            if (ModelState.IsValid)
+            if (Request.IsAuthenticated)
             {
-                //newUser.active = true;
-                newUser.created_on = DateTime.Today;
-                newUser.created_by = "Admin";
-
-                db.USERs.Add(newUser);
-                db.SaveChanges();
-                if (newUser.user_type == "Volunteer")
+                if (ModelState.IsValid)
                 {
-                    return RedirectToAction("Index");
+                    //newUser.active = true;
+                    newUser.created_on = DateTime.Today;
+                    newUser.created_by = "Admin";
+
+                    db.USERs.Add(newUser);
+                    db.SaveChanges();
+                    if (newUser.user_type == "Volunteer")
+                    {
+                        return RedirectToAction("Index");
+
+                    }
+                    else
+                        TempData["userID"] = newUser.userID;
+                    return RedirectToAction("Create", "PAID_STAFF");
 
                 }
-                else
-                    TempData["userID"] = newUser.userID;
-                return RedirectToAction("Create", "PAID_STAFF");
 
+                ViewBag.location = new SelectList(db.LOCATIONs, "locationID", "name", newUser.location);
+                ViewBag.userID = new SelectList(db.PAID_STAFF, "emplID", "pay_schedule", newUser.userID);
+                ViewBag.userID = new SelectList(db.UNIT_DIRECTOR, "emplID", "emplID", newUser.userID);
+                ViewBag.manager = new SelectList(db.USERs, "userID", "fname", newUser.manager);
+                ViewBag.userID = new SelectList(db.VOLUNTEERs, "volID", "volID", newUser.userID);
+                return View(newUser);
             }
-
-            ViewBag.location = new SelectList(db.LOCATIONs, "locationID", "name", newUser.location);
-            ViewBag.userID = new SelectList(db.PAID_STAFF, "emplID", "pay_schedule", newUser.userID);
-            ViewBag.userID = new SelectList(db.UNIT_DIRECTOR, "emplID", "emplID", newUser.userID);
-            ViewBag.manager = new SelectList(db.USERs, "userID", "fname", newUser.manager);
-            ViewBag.userID = new SelectList(db.VOLUNTEERs, "volID", "volID", newUser.userID);
-            return View(newUser);
+            else
+            {
+                return RedirectToAction("Index", "Home");
+            }
         }
 
         // GET: Users/Edit/5
         public ActionResult Edit(int? id)
         {
-            if (id == null)
+            if (Request.IsAuthenticated)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                USER uSER = db.USERs.Find(id);
+                if (uSER == null)
+                {
+                    return HttpNotFound();
+                }
+                ViewBag.location = new SelectList(db.LOCATIONs, "locationID", "name", uSER.location);
+                ViewBag.userID = new SelectList(db.PAID_STAFF, "emplID", "pay_schedule", uSER.userID);
+                ViewBag.userID = new SelectList(db.UNIT_DIRECTOR, "emplID", "emplID", uSER.userID);
+                ViewBag.manager = new SelectList(db.USERs, "userID", "fname", uSER.manager);
+                ViewBag.userID = new SelectList(db.VOLUNTEERs, "volID", "volID", uSER.userID);
+                return View(uSER);
             }
-            USER uSER = db.USERs.Find(id);
-            if (uSER == null)
+            else
             {
-                return HttpNotFound();
+                return RedirectToAction("Index", "Home");
             }
-            ViewBag.location = new SelectList(db.LOCATIONs, "locationID", "name", uSER.location);
-            ViewBag.userID = new SelectList(db.PAID_STAFF, "emplID", "pay_schedule", uSER.userID);
-            ViewBag.userID = new SelectList(db.UNIT_DIRECTOR, "emplID", "emplID", uSER.userID);
-            ViewBag.manager = new SelectList(db.USERs, "userID", "fname", uSER.manager);
-            ViewBag.userID = new SelectList(db.VOLUNTEERs, "volID", "volID", uSER.userID);
-            return View(uSER);
         }
 
         // POST: Users/Edit/5
@@ -108,33 +143,47 @@ namespace bgce_timetracker.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "userID,fname,lname,active,start_date,created_on,created_by,updated_on,updated_by,manager,location,username,email,passwrd,passwrd_last_set,passwrd_expired,passwd_salt,is_administrator,user_type,total_hours_worked")] USER uSER)
         {
-            if (ModelState.IsValid)
+            if (Request.IsAuthenticated)
             {
-                db.Entry(uSER).State = EntityState.Modified;
-                db.SaveChanges();
-                return RedirectToAction("Index");
+                if (ModelState.IsValid)
+                {
+                    db.Entry(uSER).State = EntityState.Modified;
+                    db.SaveChanges();
+                    return RedirectToAction("Index");
+                }
+                ViewBag.location = new SelectList(db.LOCATIONs, "locationID", "name", uSER.location);
+                ViewBag.userID = new SelectList(db.PAID_STAFF, "emplID", "pay_schedule", uSER.userID);
+                ViewBag.userID = new SelectList(db.UNIT_DIRECTOR, "emplID", "emplID", uSER.userID);
+                ViewBag.manager = new SelectList(db.USERs, "userID", "fname", uSER.manager);
+                ViewBag.userID = new SelectList(db.VOLUNTEERs, "volID", "volID", uSER.userID);
+                return View(uSER);
             }
-            ViewBag.location = new SelectList(db.LOCATIONs, "locationID", "name", uSER.location);
-            ViewBag.userID = new SelectList(db.PAID_STAFF, "emplID", "pay_schedule", uSER.userID);
-            ViewBag.userID = new SelectList(db.UNIT_DIRECTOR, "emplID", "emplID", uSER.userID);
-            ViewBag.manager = new SelectList(db.USERs, "userID", "fname", uSER.manager);
-            ViewBag.userID = new SelectList(db.VOLUNTEERs, "volID", "volID", uSER.userID);
-            return View(uSER);
+            else
+            {
+                return RedirectToAction("Index", "Home");
+            }
         }
 
         // GET: Users/Delete/5
         public ActionResult Delete(int? id)
         {
-            if (id == null)
+            if (Request.IsAuthenticated)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                USER uSER = db.USERs.Find(id);
+                if (uSER == null)
+                {
+                    return HttpNotFound();
+                }
+                return View(uSER);
             }
-            USER uSER = db.USERs.Find(id);
-            if (uSER == null)
+            else
             {
-                return HttpNotFound();
+                return RedirectToAction("Index", "Home");
             }
-            return View(uSER);
         }
 
         // POST: Users/Delete/5
@@ -142,10 +191,17 @@ namespace bgce_timetracker.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
-            USER uSER = db.USERs.Find(id);
-            db.USERs.Remove(uSER);
-            db.SaveChanges();
-            return RedirectToAction("Index");
+            if (Request.IsAuthenticated)
+            {
+                USER uSER = db.USERs.Find(id);
+                db.USERs.Remove(uSER);
+                db.SaveChanges();
+                return RedirectToAction("Index");
+            }
+            else
+            {
+                return RedirectToAction("Index", "Home");
+            }
         }
 
         protected override void Dispose(bool disposing)
